move test data to a dedicated file
